refactor(app): codebase structure
