+++ example.env
@@ -1,3 +1,3 @@
-PORT=3000
+PORT=8080
 CLIENT_ENDPOINT=http://localhost:5173
-VITE_WEBSOCKET_ENDPOINT=ws://localhost:3000/
+VITE_WEBSOCKET_ENDPOINT=ws://localhost:${PORT}/
