+++ src/components/InputBox/InputBox.styled.tsx
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+export const InputContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 0 20px;
+  width: 100%;
+  border-radius: 26px;
+  background-color: var(--input-bg-color);
+  box-shadow: var(--input-shadow);
+`;
