--- src/components/index.ts
@@ -1,10 +1,0 @@
-export * from './Layout';
-export * from './Avatar';
-export * from './Message';
-export * from './MessageGroup';
-export * from './ChatBox';
-export * from './Input';
-export * from './SendButton';
-export * from './Modal';
-export * from './InputBox';
-export * from './StartButton';
