+++ src/hooks/useIsImageAttachment.ts
@@ -1,3 +1,5 @@
-export const useIsImageAttachment = (type: string) => {
+const useIsImageAttachment = (type: string) => {
   return /^image\/.*/i.test(type);
 };
+
+export default useIsImageAttachment;
