+++ src/icons/CloseIcon.tsx
@@ -1,0 +1,17 @@
+type CloseIconProps = {
+  className?: string;
+};
+
+const CloseIcon = ({ className }: CloseIconProps) => (
+  <svg
+    className={className}
+    xmlns="http://www.w3.org/2000/svg"
+    height="24"
+    viewBox="0 96 960 960"
+    width="24"
+  >
+    <path d="m336 776 144-144 144 144 56-56-144-144 144-144-56-56-144 144-144-144-56 56 144 144-144 144 56 56Zm144 200q-83 0-156-31.5T197 859q-54-54-85.5-127T80 576q0-83 31.5-156T197 293q54-54 127-85.5T480 176q83 0 156 31.5T763 293q54 54 85.5 127T880 576q0 83-31.5 156T763 859q-54 54-127 85.5T480 976Z"></path>
+  </svg>
+);
+
+export default CloseIcon;
