--- src/styles/index.ts
@@ -1,2 +1,0 @@
-export * from './reset';
-export * from './root';
