+++ src/utils/genId.ts
@@ -1,0 +1,5 @@
+import { v4 } from 'uuid';
+
+const genId = () => v4();
+
+export default genId;
